move new question tests to own file
